add group ids in user login api
