+++ pkg/microservice/user/core/service/permission/user_group.go
@@ -25,6 +25,7 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/koderover/zadig/v2/pkg/config"
+	userconfig "github.com/koderover/zadig/v2/pkg/microservice/user/config"
 	"github.com/koderover/zadig/v2/pkg/microservice/user/core/repository"
 	"github.com/koderover/zadig/v2/pkg/microservice/user/core/repository/models"
 	"github.com/koderover/zadig/v2/pkg/microservice/user/core/repository/orm"
@@ -32,63 +33,6 @@ import (
 	"github.com/koderover/zadig/v2/pkg/tool/cache"
 	"github.com/koderover/zadig/v2/pkg/types"
 )
-
-const (
-	UserGroupCacheKeyFormat = "user_group_%s"
-)
-
-var allUserGroupID string
-
-// GetUserGroupByUID list all group IDs the given user with [uid] with cache
-func GetUserGroupByUID(uid string) ([]string, error) {
-	userGroupKey := fmt.Sprintf(UserGroupCacheKeyFormat, uid)
-	userCache := cache.NewRedisCache(config.RedisCommonCacheTokenDB())
-
-	// check if the cache has been set
-	exists, err := userCache.Exists(userGroupKey)
-	if err == nil && exists {
-		resp, err2 := userCache.ListSetMembers(userGroupKey)
-		if err2 == nil {
-			// if we got the data from cache, simply return it
-			return resp, nil
-		}
-	}
-
-	// if no cache is set or anything wrong happens to the cache request, search for the database
-	groups, err := orm.ListUserGroupByUID(uid, repository.DB)
-	if err != nil {
-		log.Errorf("failed to list groups by uid: %s from database, error: %s", uid, err)
-		return nil, fmt.Errorf("failed to list groups by uid: %s from database, error: %s", uid, err)
-	}
-
-	resp := make([]string, 0)
-	for _, group := range groups {
-		resp = append(resp, group.GroupID)
-	}
-
-	err = userCache.AddElementsToSet(userGroupKey, resp, setting.CacheExpireTime)
-	if err != nil {
-		// nothing should be returned since setting data into cache does not affect final result
-		log.Warnf("failed to add group IDs into user cache, error: %s", err)
-	}
-
-	return resp, nil
-}
-
-func GetAllUserGroup() (string, error) {
-	if allUserGroupID != "" {
-		return allUserGroupID, nil
-	}
-
-	group, err := orm.GetAllUserGroup(repository.DB)
-	if err != nil {
-		log.Errorf("failed to get all-user group, error: %s", err)
-		return "", err
-	}
-
-	allUserGroupID = group.GroupID
-	return allUserGroupID, nil
-}
 
 func CreateUserGroup(groupName, desc string, uids []string, logger *zap.SugaredLogger) error {
 	tx := repository.DB.Begin()
@@ -270,7 +214,7 @@ func BulkAddUserToUserGroup(groupID string, uids []string, logger *zap.SugaredLo
 	userCache := cache.NewRedisCache(config.RedisCommonCacheTokenDB())
 
 	for _, uid := range uids {
-		userGroupKey := fmt.Sprintf(UserGroupCacheKeyFormat, uid)
+		userGroupKey := fmt.Sprintf(userconfig.UserGroupCacheKeyFormat, uid)
 		err := userCache.Delete(userGroupKey)
 		if err != nil {
 			log.Warnf("failed to flush uid: %s's group id cache, error: %s", uid, err)
@@ -293,7 +237,7 @@ func BulkRemoveUserFromUserGroup(groupID string, uids []string, logger *zap.Suga
 	userCache := cache.NewRedisCache(config.RedisCommonCacheTokenDB())
 
 	for _, uid := range uids {
-		userGroupKey := fmt.Sprintf(UserGroupCacheKeyFormat, uid)
+		userGroupKey := fmt.Sprintf(userconfig.UserGroupCacheKeyFormat, uid)
 		err := userCache.Delete(userGroupKey)
 		if err != nil {
 			log.Warnf("failed to flush uid: %s 's group id cache, error: %s", uid, err)
